+++ tomcat/src/main/java/org/apache/catalina/mapping/StaticControllerMapping.java
@@ -1,4 +1,4 @@
-package com.mapping;
+package org.apache.catalina.mapping;
 
 import com.controller.Controller;
 import com.controller.StaticResourcesController;
