+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Cookie.java
@@ -16,14 +16,9 @@ public class Http11Cookie {
         extractCookieHeader(cookieHeader);
     }
 
-    public String get(final String cookieKey) {
-        return cookies.get(cookieKey);
-    }
-
     private boolean validateCookieHeader(final String cookieHeader) {
         return cookieHeader == null || cookieHeader.isEmpty();
     }
-    // yummy_cookie=choco; tasty_cookie=strawberry; JSESSIONID=656cef62-e3c4-40bc-a8df-94732920ed46
 
     private void extractCookieHeader(final String cookieHeader) {
         Arrays.stream(cookieHeader.split(";"))
@@ -31,5 +26,9 @@ public class Http11Cookie {
                 .map(s -> s.split("=", 2))
                 .filter(kv -> kv.length == 2)
                 .forEach(kv -> cookies.put(kv[0], kv[1]));
+    }
+
+    public String get(final String cookieKey) {
+        return cookies.get(cookieKey);
     }
 }
