+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Request.java
@@ -1,11 +1,5 @@
 package org.apache.coyote.http11;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 public class Http11Request {
@@ -14,7 +8,7 @@ public class Http11Request {
     private final Http11RequestHeaders headers;
     private final Http11RequestBody body;
 
-    private Http11Request(
+    public Http11Request(
             final StartLine startLine,
             final Http11RequestHeaders headers,
             final Http11RequestBody body
@@ -22,28 +16,6 @@ public class Http11Request {
         this.startLine = startLine;
         this.headers = headers;
         this.body = body;
-    }
-
-    public static Http11Request from(InputStream inputStream) throws IOException {
-        BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
-        List<String> lines = extractRequestHeadersWithStartLine(reader);
-
-        StartLine startLine = StartLine.extractStartLine(lines.getFirst());
-        Http11RequestHeaders headers = Http11RequestHeaders.extractHeaders(lines);
-        Http11RequestBody body = Http11RequestBody.from(reader, headers);
-
-        return new Http11Request(startLine, headers, body);
-    }
-
-    private static List<String> extractRequestHeadersWithStartLine(final BufferedReader bufferedReader)
-            throws IOException {
-        List<String> requestHeaders = new ArrayList<>();
-        String requestLine;
-        while ((requestLine = bufferedReader.readLine()) != null && !requestLine.isEmpty()) {
-            requestHeaders.add(requestLine);
-        }
-
-        return requestHeaders;
     }
 
     public String getSessionId() {
